refactor: rename CodeGenerator to MetaDataCollector and clear project
