+++ bro-compiler/src/main/resources/META-INF/gradle/incremental.annotation.processors
@@ -1,1 +1,1 @@
-me.xx2bab.bro.compiler.BroCompilerProcessor,AGGREGATING
+me.xx2bab.bro.compiler.BroAnnotationProcessor,AGGREGATING
