+++ bro-compiler/src/main/resources/META-INF/services/javax.annotation.processing.Processor
@@ -1,1 +1,1 @@
-me.xx2bab.bro.compiler.BroCompilerProcessor
+me.xx2bab.bro.compiler.BroAnnotationProcessor
